requirements.txt
python-dotenv
google-genai==0.7.0